cast value to int

+++ app/Services/MarketDataProviders/CoinGecko.php
@@ -109,7 +109,7 @@ final class CoinGecko extends AbstractMarketDataProvider
         return $this->isAcceptableResponse(
             $data,
             'coingecko_response_error',
-            config('explorer.coingecko_exception_frequency', 60),
+            (int) config('explorer.coingecko_exception_frequency', 60),
             'Too many empty CoinGecko responses',
             fn ($data) => Arr::get($data, 'status.error_code') !== null,
         );
@@ -120,7 +120,7 @@ final class CoinGecko extends AbstractMarketDataProvider
         return $this->isAcceptableResponse(
             $data,
             'coingecko_response_throttled',
-            config('explorer.coingecko_exception_frequency', 60),
+            (int) config('explorer.coingecko_exception_frequency', 60),
             'CoinGecko requests are being throttled',
             fn ($data) => Arr::get($data, 'status.error_code') !== null,
         );
